apps/api/pyproject.toml
[project]
name = "radd"
version = "0.1.0"
description = "Arabic-first customer service automation for Saudi e-commerce"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Cache / queue
    "redis[hiredis]>=5.2.0",
    # Vector DB
    "qdrant-client>=1.12.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Config
    "pydantic-settings>=2.6.0",
    # HTTP client
    "httpx>=0.28.0",
    # OpenAI
    "openai>=1.57.0",
    # NLP
    "transformers>=4.47.0",
    "torch>=2.5.0",
    "camel-tools>=1.5.4",
    # Utilities
    "python-multipart>=0.0.12",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
